backend/requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-jalali==4.1.2
django-js-asset==1.2.2
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
jdatetime==3.6.2
Jinja2==2.11.3
jsonschema==3.2.0
Markdown==3.3.4
MarkupSafe==1.1.1
packaging==20.9
phonenumbers==8.12.18
phonenumberslite==8.12.18
Pillow==8.1.0
PyJWT==2.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.16.12
six==1.15.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
uritemplate==3.0.1
urllib3==1.26.3
gunicorn==20.0.4